requirements.txt
anthropic>=0.40.0
finnhub-python>=2.4.0
praw>=7.7.0
sendgrid>=6.11.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
